requirement.txt
# LangChain + integrations
langchain==1.0.7
langchain-huggingface==1.0.1

# Core ML infra
transformers==4.57.1
huggingface-hub==0.36.0
sentence-transformers==2.3.0

# utilities and tools
python-dotenv
streamlit
numpy
scikit-learn